backend/app/services/imagen_logo.py
import google.generativeai as genai
from app.schemas.logo import LogoGenerateRequest, LogoConcept
from app.config import settings

# Note: Imagen integration usually requires specific model names/access
# Using a generic genai flow for illustration as per Gemini standard

async def generate_logo_concepts(request: LogoGenerateRequest) -> list[LogoConcept]:
    # In a real implementation, this would call the Imagen model
    # For now, we simulate multiple concepts
    return [
        LogoConcept(
            id="1",
            url=f"https://placehold.co/512x512?text={request.company_name}+Logo+1",
            description=f"A {request.style_preference} logo for {request.company_name} in {request.industry}."
        ),
        LogoConcept(
            id="2",
            url=f"https://placehold.co/512x512?text={request.company_name}+Logo+2",
            description=f"An alternative concept focusing on the slogan: {request.slogan or 'Premium Choice'}."
        )
    ]
