backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
google-generativeai==0.5.0
google-auth==2.29.0
python-docx==1.1.0
reportlab==4.1.0
cairosvg==2.7.1
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
